+++ packages/auth/__tests__/__system__/cli/__scripts__/auth_login_apiml_create_timeout.sh
@@ -1,0 +1,12 @@
+#!/bin/bash
+set -e
+
+HOST=$1
+PORT=$2
+USER=$3
+PASSWORD=$4
+REJECT=$5
+
+zowe auth login apiml --host $HOST --port $PORT --user $USER --password $PASSWORD --ru $REJECT
+
+exit $?