+++ packages/imperative.ts
@@ -68,14 +68,13 @@ const config: IImperativeConfig = {
                     optionDefinition: Constants.BASE_OPTION_REJECT_UNAUTHORIZED
                 },
                 tokenType: {
-                    type: "string"
+                    type: "string",
+                    optionDefinition: Constants.BASE_OPTION_TOKEN_TYPE
                 },
                 tokenValue: {
                     type: "string",
-                    secure: true
-                },
-                serviceIds: {
-                    type: "object"
+                    secure: true,
+                    optionDefinition: Constants.BASE_OPTION_TOKEN_VALUE
                 }
             },
             required: []
@@ -98,13 +97,41 @@ const config: IImperativeConfig = {
             {
                 options: "base4 --reject-unauthorized false",
                 description: "Create a zosmf profile called 'base4' to connect to default port 443 and allow self-signed certificates, " +
-                "not specifying a username, password, or host so they are not stored on disk; these will need to be specified on every command"
+                    "not specifying a username, password, or host so they are not stored on disk; these will need to be specified on every command"
             }
         ],
         updateProfileExamples: [
             {
                 options: "base1 --user newuser --password newp4ss",
                 description: "Update a base profile named 'base1' with a new username and password"
+            }
+        ],
+        authConfig: [
+            {
+                serviceName: "apiml",
+                handler: __dirname + "/auth/src/cli/ApimlAuthHandler",
+                login: {
+                    description: Constants.APIML_LOGIN_DESCRIPTION,
+                    examples: [Constants.APIML_LOGIN_EXAMPLE],
+                    options: [
+                        Constants.BASE_OPTION_HOST,
+                        Constants.BASE_OPTION_PORT,
+                        Constants.BASE_OPTION_USER,
+                        Constants.BASE_OPTION_PASSWORD,
+                        Constants.BASE_OPTION_REJECT_UNAUTHORIZED
+                    ]
+                },
+                logout: {
+                    description: Constants.APIML_LOGOUT_DESCRIPTION,
+                    examples: [Constants.APIML_LOGOUT_EXAMPLE],
+                    options: [
+                        Constants.BASE_OPTION_HOST,
+                        Constants.BASE_OPTION_PORT,
+                        Constants.BASE_OPTION_TOKEN_TYPE,
+                        Constants.BASE_OPTION_TOKEN_VALUE,
+                        Constants.BASE_OPTION_REJECT_UNAUTHORIZED
+                    ]
+                }
             }
         ]
     },
@@ -141,15 +168,6 @@ const config: IImperativeConfig = {
                     basePath: {
                         type: "string",
                         optionDefinition: ZosmfSession.ZOSMF_OPTION_BASE_PATH
-                    },
-                    tokenType: {
-                        type: "string",
-                        optionDefinition: ZosmfSession.ZOSMF_OPTION_TOKEN_TYPE
-                    },
-                    tokenValue: {
-                        type: "string",
-                        optionDefinition: ZosmfSession.ZOSMF_OPTION_TOKEN_VALUE,
-                        secure: true
                     }
                 },
                 required: []