+++ packages/zosfiles/__tests__/__integration__/cli/hMigrate/ds/cli.files.hMigrate.ds.integration.test.ts
@@ -38,6 +38,14 @@ describe("Migrate Data Set", () => {
         expect(response.stdout.toString()).toMatchSnapshot();
     });
 
+    it("should display the help in json format", async () => {
+        const response = runCliScript(__dirname + "/__scripts__/migrate_data_set_help_rfj.sh",
+            TEST_ENVIRONMENT);
+        expect(response.status).toBe(0);
+        expect(response.stderr.toString()).toBe("");
+        expect(response.stdout.toString()).toMatchSnapshot();
+    });
+
     describe("Expected failures", () => {
         it("should fail migrating a data set due to missing data set name", async () => {
             const response = runCliScript(__dirname + "/__scripts__/command/command_migrate_data_set.sh",