+++ packages/zostso/__tests__/cli/start/address-space/AddressSpace.handler.unit.test.ts
@@ -58,6 +58,7 @@ const DEFAULT_PARAMTERS: IHandlerParameters = {
         $0: "bright",
         _: ["zos-tso", "start", "address-space"]
     },
+    positionals: ["zos-tso", "start", "address-space"],
     response: {
         data: {
             setMessage: jest.fn((setMsgArgs) => {